css/css-display/reading-order/grid-order-with-shadow-dom.tentative.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Display: reading-order-items with value grid-order in Shadow DOM</title>
<link rel="help" href="https://drafts.csswg.org/css-display-4/#reading-order-items">
<link rel="author" title="Di Zhang" href="mailto:dizhangg@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="resources/focus-utils.js"></script>

<style>
.wrapper, :host {
  display: grid;
  reading-order-items: grid-order;
}
</style>

<div id=div1 tabindex=0>one</div>
<span>
  <template shadowrootmode="open" shadowrootdelegatesfocus>
    <slot name=myslot></slot>
  </template>

  <slot class="wrapper" slot=myslot>
    <div id=div2 tabindex=0 style="order: 3">two</div>
    <div id=div3 tabindex=0>three</div>
    <div id=div4 tabindex=0>four</div>
  </slot>
</span>
<div id=div5 tabindex=0>five</div>

<script>

promise_test(async () => {
  div1.focus();
  assert_equals(document.activeElement, div1);

  await navigateFocusForward();
  assert_equals(document.activeElement, div3);
  await navigateFocusForward();
  assert_equals(document.activeElement, div4);
  await navigateFocusForward();
  assert_equals(document.activeElement, div2);
  await navigateFocusForward();
  assert_equals(document.activeElement, div5);
}, `Verifies that focus order goes in flat tree order with buttons inside nested slots which have a mixture of assigned and unassigned states.`);

</script>
